requirements.txt
Flask>=2.2
requests>=2.31
python-dotenv>=1.0
